+++ noodles-bam/src/reader/records.rs
@@ -1,8 +1,9 @@
 use std::io::{self, Read};
 
-use crate::Record;
+use noodles_sam::reader::record::Fields;
 
 use super::Reader;
+use crate::Record;
 
 /// An iterator over records of a BAM reader.
 ///
@@ -13,16 +14,18 @@ where
 {
     reader: &'a mut Reader<R>,
     record: Record,
+    fields: Fields,
 }
 
 impl<'a, R> Records<'a, R>
 where
     R: Read,
 {
-    pub(crate) fn new(reader: &'a mut Reader<R>) -> Records<'_, R> {
+    pub(crate) fn new(reader: &'a mut Reader<R>, fields: Fields) -> Records<'_, R> {
         Self {
             reader,
             record: Record::default(),
+            fields,
         }
     }
 }
@@ -34,7 +37,10 @@ where
     type Item = io::Result<Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match self.reader.read_record(&mut self.record) {
+        match self
+            .reader
+            .read_record_with_fields(&mut self.record, self.fields)
+        {
             Ok(0) => None,
             Ok(_) => Some(Ok(self.record.clone())),
             Err(e) => Some(Err(e)),