+++ noodles-bam/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-  * bam/reader: Add `Reader::read_record_with_fields` to read a record and
-    decode a subset of fields.
+  * bam/reader: Add `Reader::read_record_with_fields` and
+    `Reader::records_with_fields` to read records and only decode a subset of
+    fields.
 
 ## 0.18.0 - 2022-04-14